+++ src/lib/util.ts
@@ -99,7 +99,7 @@ export function add_indend(text, indent) {
   return `<div class="indented_text" style="${style}">${text}</div>`
 }
 
-export function wrap_long_text(text, indent = false) {
+export function wrap_long_text(text = null, indent = null) {
   if (text === undefined || text === null || text === "")
     return `<div class="long_text_empty"></div>`
   if (indent) text = add_indend(text, indent)