+++ public/data/db/__table__.json.js
@@ -2,11 +2,11 @@ jsonjs.data['__table__'] =
 [
   {
     "name": "config",
-    "last_modif": 1747924819
+    "last_modif": 1757605298
   },
   {
     "name": "dataset",
-    "last_modif": 1753608552
+    "last_modif": 1757582432
   },
   {
     "name": "doc",
@@ -14,7 +14,7 @@ jsonjs.data['__table__'] =
   },
   {
     "name": "evolution",
-    "last_modif": 1757018090
+    "last_modif": 1757605298
   },
   {
     "name": "folder",
@@ -46,6 +46,6 @@ jsonjs.data['__table__'] =
   },
   {
     "name": "__table__",
-    "last_modif": 1757018090
+    "last_modif": 1757605299
   }
 ]