+++ public/data/db/config.json.js
@@ -5,14 +5,6 @@ jsonjs.data['config'] =
     "value": "contact@datannur.com"
   },
   {
-    "id": "alias_1",
-    "value": "institution : owner"
-  },
-  {
-    "id": "alias_2",
-    "value": "institution : manager"
-  },
-  {
     "id": "filter_1",
     "value": "open_data : Open Data"
   },