+++ test/Time.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest"
-import { get_time_ago } from "@js/Time"
+import { get_time_ago } from "@lib/Time"
 
 const same_date_multiple_times = [
   [new Date(2023, 10, 25, 0, 30, 0)],