+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add : filter_type "select" to column meta_localisation
 - change : increase update_app.py compatibility to python 3.8+
 - fix : datetime and wrap_long_text function with optional param null
+- remove : optional db_key param in default config jsonjsdb_config.html
 
 ## 0.12.5 (2025-09-09)